Avoid read/write collisions #2 (#25)
* Making c3.php create individual reports per request, and merging them in the end (when reporting)

c3.php originally read a the serialized file in the start of each request, and then appended to
it before saving it in the end of the request. This leads to data loss and/or corruption as soon
as you start having any amount of concurrent requests. This solution can handle any number
of concurrent requests (by merging the individual requests when they are all over), but it
requires you to collect the report from c3.php as well (since Codeception itself seems to
expect a single .serialized-file, not multiple). This is probably not ready to merge yet, but
it's enough to start a discussion about the problem.

* Reverted most of the last commit and fixed it in a better way

This way doesn't split the serialized files into one for each request, so
it works along side the report generator for local code coverage in
codeception (so you don't have to force using the c3 one). This way
the different request code coverage reports are merged at the end of
a request, and file locking makes sure multiple requests don't do so
at the same time (by moving the time of read and write closer together
and by locking the file during that time)